Corrects warnings in the CSW, CPC DSK, ZX8081 data encoding, and PRG and binary cartridges.

+++ Storage/Data/ZX8081.cpp
@@ -27,7 +27,7 @@ static std::shared_ptr<File> ZX80FileFromData(const std::vector<uint8_t> &data) 
 	uint16_t display_address = short_at(0xc, data);
 
 	// check that the end of file is contained within the supplied data
-	if(end_of_file - 0x4000 > data.size()) return nullptr;
+	if(static_cast<size_t>(end_of_file - 0x4000) > data.size()) return nullptr;
 
 	// check for the proper ordering of buffers
 	if(vars > end_of_file) return nullptr;
